app/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
psycopg2==2.9.10
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
flask-cors==3.0.10
python-dotenv==1.0.1
pytest==8.2.0
pytest-cov==4.1.0
bcrypt==4.1.2